+++ src/components/services/content/ServiceWebview.tsx
@@ -40,7 +40,7 @@ class ServiceWebview extends Component<IProps> {
           });
           this.webview.view.addEventListener('did-navigate', () => {
             if (this.props.service._webview) {
-              document.title = `Ferdium - ${this.props.service.name} ${
+              document.title = `Aitalk - ${this.props.service.name} ${
                 this.props.service.dialogTitle
                   ? ` - ${this.props.service.dialogTitle}`
                   : ''
@@ -78,7 +78,7 @@ class ServiceWebview extends Component<IProps> {
       webview.view.blur();
       webview.view.focus();
       window.setTimeout(() => {
-        document.title = `Ferdium - ${this.props.service.name} ${
+        document.title = `Aitalk - ${this.props.service.name} ${
           this.props.service.dialogTitle
             ? ` - ${this.props.service.dialogTitle}`
             : ''