+++ src/components/services/content/Services.tsx
@@ -24,16 +24,16 @@ const messages = defineMessages({
   },
   login: {
     id: 'services.login',
-    defaultMessage: 'Please login to use Ferdium.',
+    defaultMessage: 'Please login to use Aitalk.',
   },
   serverless: {
     id: 'services.serverless',
-    defaultMessage: 'Use Ferdium without an Account',
+    defaultMessage: 'Use Aitalk without an Account',
   },
   serverInfo: {
     id: 'services.serverInfo',
     defaultMessage:
-      'Optionally, you can change your Ferdium server by clicking the cog in the bottom left corner. If you are switching over (from one of the hosted servers) to using Ferdium without an account, please be informed that you can export your data from that server and subsequently import it using the Help menu to resurrect all your workspaces and configured services!',
+      'Optionally, you can change your Aitalk server by clicking the cog in the bottom left corner. If you are switching over (from one of the hosted servers) to using Aitalk without an account, please be informed that you can export your data from that server and subsequently import it using the Help menu to resurrect all your workspaces and configured services!',
   },
 });