+++ src/components/settings/navigation/SettingsNavigation.tsx
@@ -41,7 +41,7 @@ const messages = defineMessages({
   },
   supportFerdium: {
     id: 'settings.navigation.supportFerdium',
-    defaultMessage: 'About Ferdium',
+    defaultMessage: 'About Aitalk',
   },
   logout: {
     id: 'settings.navigation.logout',