+++ src/helpers/serverless-helpers.ts
@@ -12,7 +12,7 @@ export default function serverlessLogin(actions) {
   // Log into local server
   // Credentials are ignored by the server but the client requires them
   actions.user.login({
-    email: 'ferdium@localhost',
+    email: 'aitalk@localhost',
     password: 'FERDIUM_',
   });
 }