+++ src/lib/Tray.ts
@@ -48,7 +48,7 @@ export default class TrayIcon {
     this.mainWindow = firstWindow;
 
     // listen to window events to be able to set correct string
-    // to tray menu ('Hide Ferdium' / 'Show Ferdium')
+    // to tray menu ('Hide Aitalk' / 'Show Aitalk')
     this.mainWindow.on('hide', () => {
       this._updateTrayMenu(null);
     });
@@ -77,12 +77,12 @@ export default class TrayIcon {
             ? getTranslatedText(
                 tray.currentLocale,
                 'tray.hideFerdium',
-                'Hide Ferdium',
+                'Hide Aitalk',
               )
             : getTranslatedText(
                 tray.currentLocale,
                 'tray.showFerdium',
-                'Show Ferdium',
+                'Show Aitalk',
               ),
         click() {
           tray._toggleWindow();
@@ -109,7 +109,7 @@ export default class TrayIcon {
         label: getTranslatedText(
           tray.currentLocale,
           'tray.quitFerdium',
-          'Quit Ferdium',
+          'Quit Aitalk',
         ),
         click() {
           app.quit();
@@ -147,7 +147,7 @@ export default class TrayIcon {
     }
 
     this.tray = new Tray(this._getAsset('tray', INDICATOR_TRAY_PLAIN));
-    this.tray.setToolTip('Ferdium');
+    this.tray.setToolTip('Aitalk');
 
     this.trayMenu = Menu.buildFromTemplate(this.trayMenuTemplate(this));
     if (isLinux) {